+++ src/main.jsx
@@ -3,8 +3,6 @@ import ReactDOM from 'react-dom/client'
 
 import { App } from './App'
 
-import './index.css'
-
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <App />
